requirements.txt
anyio==3.6.2
APScheduler==3.6.3
apturl==0.5.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
folder-color-common==0.0.88
future==0.18.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.14.0
httpx==0.23.3
idna==2.8
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.1.2
monotonic==1.5
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psutil==5.5.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
python-apt===2.0.0-ubuntu0.20.04.5-zorin1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-telegram-bot==13.7
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rfc3986==1.5.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
tornado==6.2
tzdata==2022.7
tzlocal==4.2
ufw==0.36
urllib3==1.25.8
vboxapi==1.0
wadllib==1.3.3
xkit==0.0.0
